fix(db_bench): use u16 suffix in prefix keys, guard num=0 and cache overflow
- prefill_prefix_keys: use u16 suffix (2 bytes) matching u16 prefix,
  keeping minimum key_size at 4 bytes instead of requiring 10
- main: reject --num 0 early to prevent panic in random_range(0..0)
- config: use checked_mul for cache_mb * 1024 * 1024 to prevent overflow
- readwhilewriting: document error-logging design choice

+++ tools/db_bench/src/db.rs
@@ -65,7 +65,8 @@ pub fn prefill_prefix_keys(
             }
             let mut key = Vec::with_capacity(config.key_size);
             key.extend_from_slice(&prefix_bytes);
-            let suffix_bytes = suffix.to_be_bytes();
+            // Use u16 suffix to keep minimum key size at 4 bytes (2+2).
+            let suffix_bytes = (suffix as u16).to_be_bytes();
             key.extend_from_slice(&suffix_bytes);
             key.resize(config.key_size, 0);
